pyproject.toml
[project]
name = "kaggle-cmi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "polars>=0.20.0",
    "scikit-learn>=1.3.0",
    "lightgbm>=4.0.0",
    "scipy>=1.11.0",
    "pyyaml>=6.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "joblib>=1.3.0",
    "pyarrow>=21.0.0",
]
